+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 2.2.0-alpha.1
+
+### Fixes
+* [SDK-151] cannot-read-property-authtoken-of-undefined by @lposen in https://github.com/Iterable/react-native-sdk/pull/779
+* [SDK-181] update ios/android sdk versions by @lposen in https://github.com/Iterable/react-native-sdk/pull/780
+* [SDK-149] add-logout-functionality by @lposen in https://github.com/Iterable/react-native-sdk/pull/781
+
+## 2.2.0-alpha.0
+
+### Updates
+- Updated Android SDK version to [3.6.1](https://github.com/Iterable/iterable-android-sdk/releases/tag/3.6.1)
+- Updated iOS SDK version to [6.6.1](https://github.com/Iterable/swift-sdk/releases/tag/6.6.1)
+- Added JWT Capabilities:
+  - Added `Iterable.authhManager`, which manages the authentication flow
+  - Added `IterableRetryBackoff` and `IterableAuthFailureReason` enums
+  - Added `onJWTError` and `retryPolicy` for control over JWT flow
+- Moved all native calls to `IterableApi.ts`
+- Added JWT example to our example app
+
+### Fixes
+- Created a standalone `IterableLogger` to avoid circular dependencies
+
 ##2.1.0
 ### Updates
 * SDK is now compatible with both New Architecture and Legacy Architecture. Fix